+++ analyzer/src/main/java/nl/basjes/parse/useragent/debug/UserAgentStringMatchMakerTester.java
@@ -221,7 +221,7 @@ public class UserAgentStringMatchMakerTester extends AbstractUserAgentAnalyzer {
                 agent.setDebug(newVerbose);
                 init = options.contains("init");
             }
-            if (expected == null || expected.size() == 0) {
+            if (expected == null || expected.isEmpty()) {
                 init = true;
                 expected = null;
             }